Correcting bit precision

+++ IMUTest/IMUTest.ino
@@ -14,7 +14,7 @@ void setup() {
 
 void loop() {
   // put your main code here, to run repeatedly:
- uint8_t array[3];
+ uint16_t array[3];
  thingy.getXYZ(array);
  Serial.print("X: ");
  Serial.print(array[0], DEC);
